feat: jpql queries and dto pattern

+++ src/main/java/com/jpa/daos/MemberDao.java
@@ -13,5 +13,5 @@ public interface MemberDao {
 
     public List<Member> getAllMembers();
 
-    public Member getMemberById(int id);
+    public Member getMemberById(Long id);
 }
